MQTT Last Will and Testament (#499)
* MQTT Last Will and Testament

Register a message with the MQTT broker to publish if the client
loses connection. Also publish online/offline status - all to a
`status` topic.

avoid global variable

* Update conn-mqtt.cpp

I am fine with merging this (sorry for the delay) but I don't see the point in this line: If connectToMqtt makes a new connection, this would actually publish the message "online" twice.  And if it was already connected, what would be the point in publishing this message again? For now I'll remove this line and merge with that change.  Re-open if you have a good argument why this additional publish would be useful.

---------

Co-authored-by: dl9rdz <dl9rdz@darc.de>